Vectors in java
A vector is a dynamic array in java that can grow or shrink in size as required unlike traditional arrays where you will have to specify a fixed size. Vectors are dynamic so they don't require a fixed size and can change their size on runtime based on stored elements. Also vectors are synchronized means that they can be accessed by multiple threads at the same time without any problem. Vectors store elements in the same order in which they are added and Vectors also support duplicate elements to be added to the array again and again.